refactor(media): sink pure media selectors out of catalog (entity-engine phase A)
Phase A of ModelDrivenEntityApi.md: move the two pure media-storage selectors
that lived in catalog/api/images.py by accident of history into the media app,
where they belong on cohesion grounds (zero catalog-domain knowledge).

- media_prefetch -> media/helpers.py (model-tier; no contract change)
- serialize_uploaded_media -> new media/selectors.py, a new tier above
  {storage,schemas} in the "Media internal layers" contract, since it composes
  both into UploadedMediaSchema and the peer-isolated helpers tier may not.

_uploaded_image_urls stays in catalog: its sole caller is extract_image_urls,
which the engine keeps domain-side and relocates in phase D.

Also tighten "Media internals do not reach into the data apps": add the
previously-omitted helpers tier alongside the new selectors tier, both covered
by the existing media.models -> provenance.models ignore. Update AppBoundaries.md
to match.

Plan doc updated to mark phase A done and relocate the rides-into-engine
substrate detail into phase C.
