+++ backend/apps/catalog/api/machine_models.py
@@ -31,9 +31,10 @@ from apps.core.schemas import (
     ValidationErrorSchema,
 )
 from apps.core.types import JsonBody
-from apps.media.helpers import all_media, primary_media
+from apps.media.helpers import all_media, media_prefetch, primary_media
 from apps.media.models import EntityMedia
 from apps.media.schemas import UploadedMediaSchema
+from apps.media.selectors import serialize_uploaded_media
 from apps.provenance.helpers import claims_prefetch
 from apps.provenance.models import ChangeSetAction
 from apps.provenance.rate_limits import (
@@ -81,8 +82,6 @@ from .images import (
     extract_image_attribution,
     extract_image_urls,
     fetch_model_media_map,
-    media_prefetch,
-    serialize_uploaded_media,
 )
 from .rich_text import describe
 from .schemas import (
