+++ backend/apps/catalog/api/manufacturers.py
@@ -20,8 +20,9 @@ from apps.core.authz.types import Activity
 from apps.core.licensing import get_minimum_display_rank
 from apps.core.models import active_status_q
 from apps.core.schemas import RateLimitErrorSchema, ValidationErrorSchema
-from apps.media.helpers import all_media
+from apps.media.helpers import all_media, media_prefetch
 from apps.media.schemas import UploadedMediaSchema
+from apps.media.selectors import serialize_uploaded_media
 from apps.provenance.helpers import claims_prefetch
 from apps.provenance.rate_limits import EDIT_RATE_LIMIT_SPEC, rate_limited
 
@@ -57,8 +58,6 @@ from .helpers import (
 from .images import (
     extract_image_urls,
     fetch_model_media_map,
-    media_prefetch,
-    serialize_uploaded_media,
 )
 from .rich_text import describe
 from .schemas import (
