+++ docs/AppBoundaries.md
@@ -42,7 +42,7 @@ The edges that carry meaning beyond "higher imports lower":
 
 Two apps sit **outside** the linear spine:
 
-- **`media`** has its own internal stack (`Media internal layers`): `constants → models → {storage|processing|schemas|helpers} → authz → api`. Its `api` tier consumes `catalog` and `provenance` (upload handlers write `media_attachment` claims), so it is effectively a top-level consumer like `kiosk`. Its lower tiers are peer-isolated from `provenance`/`citation` — they must not import them — with one sanctioned edge, `media.models → provenance.models`: `media_attachment` is a claim field, so any `MediaSupportedModel` is structurally a `ClaimControlledModel`, and the inheritance encodes that as a compile-time guarantee. The reverse direction — data apps importing `media` — is forbidden outright.
+- **`media`** has its own internal stack (`Media internal layers`): `constants → models → {storage|processing|schemas|helpers} → selectors → authz → api`. `selectors` sits above the peer group because it composes `storage` + `schemas` into wire schemas — something the peer-isolated `helpers` tier (which only reads prefetched rows off `models`) may not do. Its `api` tier consumes `catalog` and `provenance` (upload handlers write `media_attachment` claims), so it is effectively a top-level consumer like `kiosk`. Its lower tiers are peer-isolated from `provenance`/`citation` — they must not import them — with one sanctioned edge, `media.models → provenance.models`: `media_attachment` is a claim field, so any `MediaSupportedModel` is structurally a `ClaimControlledModel`, and the inheritance encodes that as a compile-time guarantee. The reverse direction — data apps importing `media` — is forbidden outright.
 - **`claim_ingest`** is the bulk write path: a management-command-driven consumer kept deliberately out of the spine so `catalog → claim_ingest` stays forbidden (sealed in both directions). It and `claim_edit` are mutually independent — the two claim-write surfaces share no code, only the provenance/core substrate they both write through.
 
 `kiosk` references `catalog` (FK to `Title`); its audit FKs use `AUTH_USER_MODEL` via Django core, not `accounts.UserProfile`. Kiosk configs are operational settings, not claims, so the app stays out of `provenance`.
